No need for separate find handler package
